requirements.txt
streamlit
streamlit-drawable-canvas
numpy
pillow
opencv-python-headless
pywavelets
scipy
graphviz
streamlit-drawable-canvas